requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.3
werkzeug==2.3.7
pyjwt==2.8.0
httpx==0.25.0
schedule==1.2.0
openai==1.3.0
asgiref==3.7.2
PyPDF2==3.0.1
Pillow>=10.4.0